changed distance check
adjusted the distance check to calculate the square of the magnitude rather than the sqrt. added a few new variables to keep track of the closest object to the player. All other objects check their distance to the player against the distance of the closest object. If their distance to the player is shorter, the closest object will change to that object. The player can only interact with the closest object.